Revert "Fixes Runtime with qdel loop + adds ability to pick up more animals"
This reverts commit 3304e5e14f4d2d95e297a754889267c29eaaa0da.

+++ code/modules/mob/living/inhand_holder.dm
@@ -5,13 +5,14 @@
 	desc = "Yell at coderbrush."
 	icon = null
 	icon_state = ""
+	flags_1 = DROPDEL_1
 	var/mob/living/held_mob
-	var/can_head = FALSE
+	var/can_head = TRUE
+	var/destroying = FALSE
 
-/obj/item/clothing/head/mob_holder/Initialize(mapload, mob/living/M, _worn_state, head_icon, lh_icon, rh_icon, _can_head_override = FALSE)
+/obj/item/clothing/head/mob_holder/Initialize(mapload, mob/living/M, _worn_state, head_icon, lh_icon, rh_icon, _can_head = TRUE)
 	. = ..()
-	if(_can_head_override)
-		can_head = _can_head_override
+	can_head = _can_head
 	if(head_icon)
 		alternate_worn_icon = head_icon
 	if(_worn_state)
@@ -25,8 +26,9 @@
 	deposit(M)
 
 /obj/item/clothing/head/mob_holder/Destroy()
+	destroying = TRUE
 	if(held_mob)
-		release()
+		release(FALSE)
 	return ..()
 
 /obj/item/clothing/head/mob_holder/proc/deposit(mob/living/L)
@@ -43,26 +45,39 @@
 /obj/item/clothing/head/mob_holder/proc/update_visuals(mob/living/L)
 	appearance = L.appearance
 
-/obj/item/clothing/head/mob_holder/dropped()
-	..()
-	release()
-
-/obj/item/clothing/head/mob_holder/proc/release(del_on_release = TRUE)//set true when not relying on DROPDEL_1
-	if(held_mob)
-		var/mob/living/m = held_mob
-		m.forceMove(get_turf(m))
-		m.reset_perspective()
-		m.setDir(SOUTH)
-		held_mob = null
+/obj/item/clothing/head/mob_holder/proc/release(del_on_release = TRUE)
+	if(!held_mob)
+		if(del_on_release && !destroying)
+			qdel(src)
+		return FALSE
 	if(isliving(loc))
 		var/mob/living/L = loc
 		L.dropItemToGround(src)
-	if(del_on_release)
+	if(held_mob)
+		held_mob.forceMove(get_turf(held_mob))
+		held_mob.reset_perspective()
+		held_mob.setDir(SOUTH)
+		held_mob = null
+	if(del_on_release && !destroying)
 		qdel(src)
+	return TRUE
 
 /obj/item/clothing/head/mob_holder/relaymove(mob/user)
-	release(TRUE)
+	release()
 
 /obj/item/clothing/head/mob_holder/container_resist()
-	release(TRUE)
+	release()
 
+/obj/item/clothing/head/mob_holder/drone/deposit(mob/living/L)
+	. = ..()
+	if(!isdrone(L))
+		qdel(src)
+	name = "drone (hiding)"
+	desc = "This drone is scared and has curled up into a ball!"
+
+/obj/item/clothing/head/mob_holder/drone/update_visuals(mob/living/L)
+	var/mob/living/simple_animal/drone/D = L
+	if(!D)
+		return ..()
+	icon = 'icons/mob/drone.dmi'
+	icon_state = "[D.visualAppearence]_hat"
